backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.0
pydantic[email]==2.11.1
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
httpx==0.28.1
celery[redis]==5.5.1
redis==5.3.0
websockets==15.0.1
anthropic==0.52.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
telethon==1.37.0
cryptg==0.4.0
